# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KOHO Fintech API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Account {
  id: number;
  name: string;
  email: string;
  spendableBalance: number;
  coverAvailable: number;
  coverLimit: number;
  accountNumber?: string;
  transitNumber?: string;
  institutionNumber?: string;
  institutionName?: string;
  notificationCount?: number;
}

export interface AccountSummary {
  totalBalance: number;
  spendable: number;
  vault: number;
  roundups: number;
  goals: number;
  interestRate: number;
  cashbackThisMonth?: number;
  cashbackAllTime?: number;
}

export type TransactionStatus = typeof TransactionStatus[keyof typeof TransactionStatus];


export const TransactionStatus = {
  completed: 'completed',
  disputed: 'disputed',
  pending: 'pending',
} as const;

export interface Transaction {
  id: number;
  merchant: string;
  amount: number;
  category: string;
  date: string;
  time: string;
  status: TransactionStatus;
  /** @nullable */
  merchantIcon?: string | null;
  /** @nullable */
  merchantColor?: string | null;
  isCredit?: boolean;
}

export interface SpendingCategory {
  category: string;
  amount: number;
  count: number;
}

export interface CreditRange {
  label: string;
  min: number;
  max: number;
  color: string;
}

export interface CreditProfile {
  score: number;
  scoreChange: number;
  status: string;
  nextRange: number;
  creditBuildingActive?: boolean;
  creditBuildingActionRequired?: boolean;
  securedCreditActive?: boolean;
  rentReportingActive?: boolean;
  ranges: CreditRange[];
}

export interface Savings {
  vault: number;
  roundups: number;
  interestRate: number;
  totalInterestEarned: number;
}

export interface Goal {
  id: number;
  name: string;
  targetAmount: number;
  currentAmount: number;
  emoji: string;
  /** @nullable */
  deadline?: string | null;
}

export interface GoalInput {
  name: string;
  targetAmount: number;
  emoji: string;
  /** @nullable */
  deadline?: string | null;
}

export interface Cover {
  available: number;
  limit: number;
  isActive: boolean;
  fee?: number;
}

export interface Cashback {
  totalEarned: number;
  earnedThisMonth: number;
  earnedLastMonth: number;
  rate: number;
}

export interface CryptoHolding {
  symbol: string;
  name: string;
  amount: number;
  value: number;
  change24h: number;
}

export interface Crypto {
  balance: number;
  allTimeChange: number;
  holdings: CryptoHolding[];
}

export interface Plan {
  name: string;
  renewalDate: string;
  interestRate: number;
  totalBalance: number;
  spendable: number;
  vault: number;
  roundups?: number;
  goals?: number;
  cashbackEarnedLastMonth?: number;
  cashbackAllTime?: number;
}

export type ListTransactionsParamsSort = typeof ListTransactionsParamsSort[keyof typeof ListTransactionsParamsSort];

export const ListTransactionsParamsSort = {
  date_desc: 'date_desc',
  date_asc: 'date_asc',
  amount_desc: 'amount_desc',
  amount_asc: 'amount_asc',
} as const;

export type ListTransactionsParams = {
search?: string;
category?: string;
dateRange?: 'this_week' | 'this_month';
limit?: number;
offset?: number;
sort?: ListTransactionsParamsSort;
};

export interface SpendingMonthSummary {
  total: number;
  categories: SpendingCategory[];
}

export interface TopMerchant {
  merchant: string;
  amount: number;
  count: number;
  /** @nullable */
  merchantColor?: string | null;
}

export interface SpendingInsights {
  currentMonth: string;
  previousMonth: string;
  thisMonth: SpendingMonthSummary;
  lastMonth: SpendingMonthSummary;
  topMerchants: TopMerchant[];
}

export interface MonthlyTrendPoint {
  month: string;
  label: string;
  total: number;
}