# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: KOHO Fintech API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List all transactions grouped by date
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: dateRange
          in: query
          schema:
            type: string
            enum: [this_week, this_month]
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
        - name: sort
          in: query
          schema:
            type: string
            enum: [date_desc, date_asc, amount_desc, amount_asc]
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
...
  /savings/goals:
    get:
      operationId: listGoals
      tags: [savings]
      summary: List savings goals
      responses:
        "200":
          description: Goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [savings]
      summary: Create a savings goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
...
components:
  schemas:
...
    Transaction:
      type: object
      required: [id, merchant, amount, category, date, time, status]
      properties:
        id:
          type: integer
        merchant:
          type: string
        amount:
          type: number
        category:
          type: string
        date:
          type: string
        time:
          type: string
        status:
          type: string
          enum: [completed, disputed, pending]
        merchantIcon:
          type: ["string", "null"]
        merchantColor:
          type: ["string", "null"]
        isCredit:
          type: boolean
...
    Goal:
      type: object
      required: [id, name, targetAmount, currentAmount, emoji]
      properties:
        id:
          type: integer
        name:
          type: string
        targetAmount:
          type: number
        currentAmount:
          type: number
        emoji:
          type: string
        deadline:
          type: ["string", "null"]
    GoalInput:
      type: object
      required: [name, targetAmount, emoji]
      properties:
        name:
          type: string
        targetAmount:
          type: number
        emoji:
          type: string
        deadline:
          type: ["string", "null"]
...