# lib/api-zod/src/generated/types/goalInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KOHO Fintech API
 * OpenAPI spec version: 0.1.0
 */

export interface GoalInput {
  name: string;
  targetAmount: number;
  emoji: string;
  /** @nullable */
  deadline?: string | null;
}